% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{ds.jobs.admin.cancel}
\alias{ds.jobs.admin.cancel}
\title{Cancel any job (admin only)}
\usage{
ds.jobs.admin.cancel(conns, job_id, admin_key)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job_id}{Character; job ID to cancel.}

\item{admin_key}{Character; the admin key matching \code{dsjobs.admin_key} or
\code{DSJOBS_ADMIN_KEY} on the server.}
}
\description{
Cancel any job (admin only)
}
