% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{ds.jobs.submit}
\alias{ds.jobs.submit}
\title{Submit a job to all nodes}
\usage{
ds.jobs.submit(conns, job)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job}{A dsjobs_job object.}
}
\value{
A dsjobs_submission with job_id and per-server details.
}
\description{
Submit a job to all nodes
}
