% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ds.hpc.scheduler_status}
\alias{ds.hpc.scheduler_status}
\title{Retired analyst scheduler status}
\usage{
ds.hpc.scheduler_status(conns)
}
\arguments{
\item{conns}{Ignored compatibility argument.}
}
\value{
This function always raises an error before contacting a server.
}
\description{
Scheduler topology is node-operator information and is no longer queried
through an analyst DataSHIELD connection.
}
