% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{ds.hpc.status}
\alias{ds.hpc.status}
\title{Get dsHPC job status}
\usage{
ds.hpc.status(conns, job_id)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job_id}{Character; opaque bearer or the symbol assigned by a domain
workflow (see \code{ds.hpc.job_id()}).}
}
\value{
A \code{dshpc_result} with one status list per site.
}
\description{
Get dsHPC job status
}
\examples{
\dontrun{
# conns <- DSI::datashield.login(...)  # live DataSHIELD session
# job <- dsImagingClient::ds.imaging.qc.metrics(conns, symbol = "jobA")
ds.hpc.status(conns, "jobA")                 # by symbol
ids <- ds.hpc.job_id(conns, "jobA")
ds.hpc.status(conns, ids[[1]])               # by opaque bearer
}
}
