% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{ds.hpc.studio_data}
\alias{ds.hpc.studio_data}
\title{Retired analyst dsHPC Studio data}
\usage{
ds.hpc.studio_data(conns, server = NULL, label = NULL, mode = "mine+global")
}
\arguments{
\item{conns}{Ignored compatibility argument.}

\item{server}{Ignored compatibility argument.}

\item{label}{Ignored compatibility argument.}

\item{mode}{Ignored compatibility argument.}
}
\value{
This function always raises an error before contacting a server.
}
\description{
Retired analyst dsHPC Studio data
}
